requirements.txt
streamlit>=1.30
pytest>=7.0
google-genai>=1.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
